requirements.txt
PyAutoGUI~=0.9.54
pystray~=0.19.5
pynput~=1.7.7
pillow~=10.3.0
pyinstaller
